+++ tests/toisto/test_app.py
@@ -18,21 +18,20 @@ class AppTest(unittest.TestCase):
         os.environ['COLUMNS'] = "120"  # Fake that the terminal is wide enough.
         self.assertRaises(SystemExit, main)
         self.assertEqual(
-            f"usage: {sys.argv[0].rsplit('/', maxsplit=1)[-1]} [-h] [-V] "
-            "[-t {topic}] [-f {topic file}] [-s {option}] {language to practice} {your language} [{command}]\n",
+            f"usage: {sys.argv[0].rsplit('/', maxsplit=1)[-1]} [-h] [-V] {{practice,progress}} ...\n",
             write.call_args_list[0][0][0]
         )
 
     @patch("os.system", Mock())
     @patch("builtins.input", Mock(side_effect=[EOFError]))
-    @patch.object(sys, "argv", ["toisto", "fi", "nl"])
+    @patch.object(sys, "argv", ["toisto", "practice", "fi", "nl"])
     def test_practice(self):
         """Test that the practice command can be invoked."""
         with patch("rich.console.Console.print") as patched_print:
             main()
         self.assertTrue(patched_print.call_args_list[0][0][0].startswith("👋 Welcome to [underline]Toisto"))
 
-    @patch.object(sys, "argv", ["toisto", "fi", "nl", "progress"])
+    @patch.object(sys, "argv", ["toisto", "progress", "fi", "nl"])
     def test_progress(self):
         """Test that the progress command can be invoked."""
         with patch("rich.console.Console.print") as patched_print: