+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to Toisto will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.6.1 - 2023-01-14
+
+### Fixed
+
+- Fix the order of quizzes so that singulars are quizzed before plurals, present tense before past tense, etc.
+
 ## v0.6.0 - 2023-01-07
 
 ### Added